Move EventHandler to subclass. Fix type mapping issues.
